requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backtesting==0.6.4
bcrypt==5.0.0
blinker==1.9.0
bokeh==3.7.3
cachetools==6.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
distro==1.9.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
google-auth==2.41.1
google-genai==1.45.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.2
MetaTrader5==5.0.4993
multidict==6.7.0
narwhals==1.40.0
numpy==2.2.6
openai==2.5.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
ta-lib @ file:///C:/A%20NEW%20BOT/Forex-Trading-Bot/ta_lib-0.6.3-cp311-cp311-win_amd64.whl
tenacity==9.1.2
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
xai-sdk==1.3.1
xyzservices==2025.4.0
yarl==1.22.0
zipp==3.23.0
